requirements.txt
anyascii==0.2.0
asgiref==3.3.4
beautifulsoup4==4.8.2
certifi==2021.5.30
chardet==4.0.0
cloudinary==1.25.0
dj-database-url==0.5.0
Django==2.2.24
django-filter==2.4.0
django-modelcluster==5.1
django-taggit==1.4.0
django-treebeard==4.5.1
djangorestframework==3.12.4
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==2.10
l18n==2020.6.1
openpyxl==3.0.7
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
telepath==0.2
typing-extensions==3.10.0.0
Unidecode==1.2.0
urllib3==1.26.5
wagtail==2.13.1
wagtailcloudinary==1.0.0
webencodings==0.5.1
whitenoise==5.3.0
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0
